web3/requirements.txt
py-solc-x
black
